[Refactor] Extract TabletColumn from tablet schema
